Refactr amplitude generation
With this commit (and several preceding) a new strategy to generate
squared matrix elemtns is introduced. The main reason is to reduce
clutter in me4qcd.frm that leads to more transparancy in the code.

The key idea is to generate amplitudes and their conjugates using
QFRAF (previously conjugation was done in me4qcd.frm). We do this
by exchangin incoming and outgoing momenta for the conjugate
amplitude that makes reasoning about momenta and frmion flows much
more easy.

BUG: uU2uUg_0_0 still does not agree with FeynCalc, however it looks
     to be something wrong with color factors. Hope to fix this soon.